+++ app/portfolio/[slug]/page.tsx
@@ -1,22 +1,22 @@
-import { allProjects } from 'contentlayer/generated';
+import { allClients } from 'contentlayer/generated';
 import { getMDXComponent } from 'next-contentlayer/hooks';
-import components from 'components/MDXComponents';
 import Image from 'next/image';
 import { notFound } from 'next/navigation';
+import components from 'components/MDXComponents';
 
-export default function Project({ params }: { params: { slug: string } }) {
-  const project = findProjectBySlug(params.slug);
+export default function Client({ params }: { params: { slug: string } }) {
+  const client = findClientBySlug(params.slug);
 
-  if (!project) {
+  if (!client) {
     notFound();
   }
 
-  const Component = getMDXComponent(project.body.code);
+  const Component = getMDXComponent(client.body.code);
 
   return (
     <article className="flex flex-col items-start justify-center w-full max-w-2xl mx-auto mb-16">
       <h1 className="mb-4 text-3xl font-bold tracking-tight text-black md:text-5xl dark:text-white">
-        {project.name}
+        {client.name}
       </h1>
       <div className="flex flex-col items-start justify-between w-full mt-2 md:flex-row md:items-center">
         <div className="flex items-center">
@@ -29,11 +29,11 @@ export default function Project({ params }: { params: { slug: string } }) {
           />
           <p className="ml-2 text-sm text-gray-700 dark:text-gray-300">
             {'Koen van Gilst / '}
-            {project.year}
+            {client.year}
           </p>
         </div>
         <p className="mt-2 text-sm text-gray-600 dark:text-gray-400 min-w-32 md:mt-0">
-          {project.readingTime.text}
+          {client.readingTime.text}
         </p>
       </div>
       <section className="w-full mt-4 prose dark:prose-dark max-w-none">
@@ -43,16 +43,15 @@ export default function Project({ params }: { params: { slug: string } }) {
   );
 }
 
-function findProjectBySlug(slug: string) {
-  return allProjects.find((project) => project.slug === slug);
+function findClientBySlug(slug: string) {
+  return allClients.find((client) => client.slug === slug);
 }
 
 export function generateMetadata({ params }: { params: { slug: string } }) {
-  const project = findProjectBySlug(params.slug);
+  const client = findClientBySlug(params.slug);
 
   return {
-    title: `${project?.name}`,
-    description: project?.summary,
-    image: `https://koenvangilst.nl${project?.image}`
+    title: client?.name,
+    description: client?.summary
   };
 }
