--- pages/portfolio/[slug].tsx
@@ -1,28 +1,0 @@
-import { useMDXComponent } from 'next-contentlayer/hooks';
-import components from 'components/MDXComponents';
-import ClientLayout from 'layouts/client';
-import { allClients } from 'contentlayer/generated';
-import type { Client } from 'contentlayer/generated';
-
-export default function Post({ client }: { client: Client }) {
-  const Component = useMDXComponent(client.body.code);
-
-  return (
-    <ClientLayout client={client}>
-      <Component components={{ ...components } as any} />
-    </ClientLayout>
-  );
-}
-
-export async function getStaticPaths() {
-  return {
-    paths: allClients.map((p) => ({ params: { slug: p.slug } })),
-    fallback: false
-  };
-}
-
-export async function getStaticProps({ params }) {
-  const client = allClients.find((client) => client.slug === params.slug);
-
-  return { props: { client } };
-}
